+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.10.0
+
+- add component gits container shadow
+- add component gits search
+- refactor date time extension
+- add extension currency
+
 ## 0.9.0
 
 - add component gits bottom sheet